Update dipify_gcode.py -  Leave probe dipped during UV
During UV exposure, the probe is now left in the resin. Also, when the probe is dipped, a brief UV pulse is given to stop resin running together on the same layer.
The "skim height" at which the probe just lifts off the surface between dots is now an actual variable.